database/rds/requirements.txt
appdirs==1.4.4
attrs==19.3.0
aws-cdk.assets==1.51.0
aws-cdk.aws-cloudwatch==1.51.0
aws-cdk.aws-ec2==1.51.0
aws-cdk.aws-efs==1.51.0
aws-cdk.aws-events==1.51.0
aws-cdk.aws-iam==1.51.0
aws-cdk.aws-kms==1.51.0
aws-cdk.aws-lambda==1.51.0
aws-cdk.aws-logs==1.51.0
aws-cdk.aws-rds==1.51.0
aws-cdk.aws-s3==1.51.0
aws-cdk.aws-s3-assets==1.51.0
aws-cdk.aws-sam==1.51.0
aws-cdk.aws-secretsmanager==1.51.0
aws-cdk.aws-sqs==1.51.0
aws-cdk.aws-ssm==1.51.0
aws-cdk.cloud-assembly-schema==1.51.0
aws-cdk.core==1.51.0
aws-cdk.cx-api==1.51.0
aws-cdk.region-info==1.51.0
black==19.10b0
cattrs==1.0.0
click==7.1.2
constructs==3.0.4
jsii==1.8.0
pathspec==0.8.0
publication==0.0.3
python-dateutil==2.8.1
# Editable Git install with no remote (rds==0.0.1)
#-e /home/rikesh/Documents/utils/aws/cdk-walkthrough/database/rds/rds
regex==2020.6.8
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
